+++ src/network_puzzles/gui/labels.kv
@@ -39,17 +39,6 @@
     text_size: self.size
     padding: dp(3)
 
-<TerminalLabel>:
-    background_color: app.theme.bg3
-    foreground_color: app.theme.fg3
-    halign: 'left'
-    valign: 'bottom'
-    font_name: 'RobotoMono-Regular'
-    font_size: sp(12)
-    text_size: self.size
-    readonly: True
-    do_wrap: False
-
 <ToolTip>:
     size_hint: (None, None)
     text_size: (None, None)
