+++ src/network_puzzles/gui/networkpuzzles.kv
@@ -12,8 +12,8 @@
 #:include network_puzzles/gui/layouts.kv
 #:include network_puzzles/gui/links.kv
 #:include network_puzzles/gui/views.kv
-#:include network_puzzles/gui/devices.kv
 #:include network_puzzles/gui/popups.kv
+#:include network_puzzles/gui/devices.kv
 
 
 ThemedBoxLayout:
@@ -100,3 +100,14 @@ ThemedBoxLayout:
             id: terminal
             size_hint_y: None
             height: self.line_height * 7
+
+<TerminalLabel>:
+    background_color: app.theme.bg3
+    foreground_color: app.theme.fg3
+    halign: 'left'
+    valign: 'bottom'
+    font_name: 'RobotoMono-Regular'
+    font_size: sp(12)
+    text_size: self.size
+    readonly: True
+    do_wrap: False
